Sync when writing the file
Glean does not attempt to do os.fsync after writing files, which
is not safe in Python since Python does not guarantee the file
being written even if it is closed on Python level.

This might result in ifup complaining with unknown interface as
the file is still in cache.

Depends-On: https://review.opendev.org/677796

Change-Id: I05fd94662c409660857d4bc66b9f6354ac21496a