Do not bring up udev assigned interfaces
The extant code is designed to loop over every device in
/sys/class/net and bringing the interface "up" to see if it valid and
something that should be configured.

As described inline, if we bring "up" an interface it can accept an RA
and get an ipv6 address assigned by the kernel.  NetworkManager will
then refuse to further configure the interface, leaving the host
generally without ipv4 networking.

The per-interface loop only actually happens on older platforms that
don't use systemd.  On systemd, glean is called for each interface
individually by udev rules.  However, we fall into the old code path,
just with one interface to work with, rather than all of them.

Thus this initial hack detects that case (by noting we were passed the
interface explicitly) and short-circuits activity check; it just
assumes that if udev asked (and it's not of a device type we don't
support), then the interface should be configured.  The interface will
*not* be put into the "up" state.

We should follow-on this change with a removal of this loop and
cleaning up the non udev/systemd activation paths.  However, this
depends on a few longer term things:

- removing Trusty support (which still hangs on by the skin of its
  teeth in OpenStack Infra, so we need to be not building nodes there)
- evaluating what Gentoo is doing in the non-systemd case.
- making sure bifrost doesnt' depend on this (likely only other user?)

In the mean time, this should fix the race conditions we've been
seeing on system+network-manager platforms.

Change-Id: I6ce51a8755e1892d3010eefd365fbad6bcec137b