write one resolv config
When systemd-resolved is enabled dns info is only written to
/etc/systemd/resolved.conf.  If systemd-resolved is not enabled dns info is
only written to /etc/resolv.conf.  We do not write to /etc/resolv.conf
if systemd-resolved is enabled because /etc/resolv.conf is a symlink pointing
to ../run/systemd/resolve/stub-resolv.conf.  This target file does not exist
before systemd-resolved is run so glean would fail with:
    'FileNotFoundError: [Errno 2] No such file or directory: '/etc/resolv.conf'

Change-Id: I644e0b50cfb7bb00a108160b99c0c1359d6a9dd4
Signed-off-by: Matthew Thode <mthode@mthode.org>