Reduce metadata read/parsing overhead
There's not need to open and parse this file multiple times.  Do it
once and pass the json decoded variable to the various functions.  The
functions are already probing for data and handling it if keys are not
there, so we pass a blank dictionary if there is no meta-data.

Change-Id: I28f5cd72004fcb1cdc5ba0a1255376da99e47140