Override NetworkManager to wait for udev-settle
As described inline, we think that NetworkManager is starting before
glean because the udev events have not been processed and the glean
services have not started.  Since there is nothing blocking it, the
network-pre.target is reached and systemd is free to start
NetworkManager.

Although it is not reccommended to wait on udev-settle because it
could result in long timeouts [1], I don't think we have a choice here
because the fundamental nature of this tool is to override settings
before NetworkManager goes off and does its own thing.

[1] https://www.freedesktop.org/software/systemd/man/systemd-udev-settle.service.html

Co-Authored-By: Dmitry Tantsur <dtantsur@protonmail.com>
Change-Id: I1c9c68d3eb5fbeb42901f2ed95860641cb2d676f