Stop requiring /usr/local/bin links for glean.sh
Since Ic16f134fe34293bb68e7c632dd320f523366320d simple-init has
installed glean into a virtualenv.  Despite this, it still needs to
setup a range of global symlinks (in /usr/local/bin) so glean can
blindly call "glean" or "glean.sh".

This means that glean installation isn't actually stand-alone.  Unless
simple-init sets up these global symlinks the glean install doesn't
work.  This makes it very annoying to try and update the way the
scripts are working, because we have to merge changes into simple-init
as well.

We can make the installation self-consistent by using the install
tool.  The init scripts call glean.sh -- we can find the full path to
glean.sh using pkg_resources and write that into the files.

glean.sh wants to call the python tool "glean".  This is slightly
harder, because at runtime the script doesn't really know how to
invoke this (i.e. in a virtualenv, it should call
/path/to/virtualenv/bin/glean).  To allow for this, write a sibling
file next to "glean.sh" to invoke glean with the interpreter the
install is running under patched in.  This way, glean.sh can call this
wrapper relative to itself and get the right thing.  Add a __main__ to
allow glean to be called like this, and update the glean.sh script.

Change-Id: I1adfecf0d5c53648ee857be502216fd8d6cc5e16