Create "legacy" script path
This updates all the init scripts that call glean and expect it to
iterate the network interfaces to call a "glean-legacy.sh" script.

This is opposed to the udev path; where glean gets called for each
interface activated by the kernel.  Follow-on changes will modify this
path to do less work for each invocation.

Change-Id: I789890b3d55838f3b70c65e519991fed0eb74e6e