write_redhat_interfaces: refactor to walk interfaces first
We currently go through each network an assume that we're writing out
one ifcfg-<interface> file for each entry we see.

This fails when we have an ipv4 and and ipv6 entry for the same
interface.  In this case, we need to combine the details into the same
ifcfg-<interface> config file.

This is a small refactor of write_redhat_interfaces to filter out the
interfaces list before we start writing the config files.  This is the
first step in reworking things so an output function can take multiple
networks to configure.

This is intended to be only a refactor and have no functional change.

Change-Id: I0aac3674fd17ff2bfcf3542e27097e171036c837