write_redhat_interfaces: pass multiple networks to output functions
This is a follow-on to I0aac3674fd17ff2bfcf3542e27097e171036c837

Prior to this we are walking the list of networks and setting them up
one-by-one.  As noted in the prior change, this fails when two
networks refer to the same interface (such as configuring ipv4 &
ipv6).

This builds a dict of interface -> [networks] mapping.  For the most
part, interfaces will only have one entry in their networks.  However,
some will have ipv4 & ipv6.

For now, we filter out the ipv6 networks; this will be addressed in a
follow-on.  This means we are effectively keeping the status-quo;
which is to ignore ipv6.  This change is intended to have no
operational changes.

Change-Id: I7eb3a4e33ae1d924a46e85035ac283a6755a74a4