_network_info: Clean up TYPE=Ethernet handling
This handling for "TYPE=Ethernet" came in with the suse split in
I20bffabd333ea290d8712ec2a467f2b2d5678f3a.

I think this is a case of over-fitting the test data incorrectly.  If
"TYPE=Ethernet" is not specified, it is assumed (e.g. [1]).  For
whatever reason this was not put in some of the test fixtures, which
have been copied over-and-over.  When the referenced change moved to
more "generic" handling it was written to fit this pattern, but really
we can/should specify the type.

This is a cleanup to help make combining ipv6 information easier in
the future.  This should have no operational effect, since this was
the default anyway.

[1] https://github.com/NetworkManager/NetworkManager/blob/6a68008e44bf2e9b6bc464825ec61bd062120f4a/src/core/settings/plugins/ifcfg-rh/nms-ifcfg-rh-reader.c#L6446

Change-Id: I09a79359bdc13973f67f0503eee9fa9fd7586ead