redhat-ish platforms: use IPV6_DEFAULTGW
It turns out that prior to NetworkManager commit [1] you can not
specify default routes in the routes[6]-<interface> file.  So
Iac0760ed1dd33c06b81cdd2ea3885279d2aab878 works OK for modern
releases, for maximum compatability we should walk the routes and put
the default ipv6 route in the ifcfg file directly, instead of in the
separate route file.

[1] https://cgit.freedesktop.org/NetworkManager/NetworkManager/commit/?id=c167e0140babcf1a045cee34ce4938f5087f8fe6&utm_source=anzwix

Change-Id: I41381bb65d980002c4ccfd2eca643ec51529360a