tests: symlink same files together
Many of the test fixture output comparision files are the same.
Updating them all when you change something can get quite annoying.

Symlink files that are expected to be the same together.  The general
pattern is to use the most "generic" version -- so centos/fedora/rocky
symlink back to the "redhat" version, "ubuntu" symlinks back to the
"debian" version, etc.

Change-Id: I6d46cc076ed8571521a1edec06d4064c424c59f5