testing: add vexxhost sample config
This adds a sample setup for vexxhost, which specifies DHCP and IPv6
SLAAC.

Honestly, I think this reveals that Debian/Ubuntu aren't handling this
combo quite right.  They don't seem to be correctly specifying the
ipv6 autoconfig options for the interface.  None of this path has been
touched in a long time, so I am inclined to leave well enough alone.
his at least provides a basis for tweaking in the future.

This has some slight reworking of things in the test generator.  The
sample configs are all using the more modern stable-name "ens3"
instead of "eth0", so we need to parameterize this to some bits of the
testing.  Also makes sure it sets up the ipv6 address as a hostname.

Change-Id: If87bb8f981cb2e7e12918919210d0d58349f17a2