redhat-ish platforms: refactor simplification of interface writing
It was pointed out in reviews of
Iac0760ed1dd33c06b81cdd2ea3885279d2aab878 (thanks clarkb) that a
single network is just a case of a multiple network list len()==1,
thus we can simplify this path.

This is a refactor to implement this, removing _write_rh_interfaces()
and combining it into write_redhat_interfaces().  I was probably a bit
deep into it when I wrote a lot of these comments, rework them a bit.

Change-Id: Iede6054d84cffa28e1365501918264a779e5db76