Use importlib when pkg_resources isn't available
pkg_resources is part of setuptools which is no longer included in
python environments by default in python3.12. Importlib.resources can be
used instead and is included in stdlib starting with python3.7.
Unfortunately the ability to extract directories isn't available until
python3.12 in importlib.resources and we rely on this functionality. We
address these problems by trying to use pkg_resources first which should
work for anything python3.11 and older and if not available use
importlib.resources which should work for python3.12 and newer.

Change-Id: I97d9150f92960f0e7ab5efb60523a67d750c7646