Update zuul config to drop xenial jobs
Glean was using the old python35 job template which depends on xenial to
run. We want to drop xenial entirely so we drop that template. While we
are at it we drop all of the templates and use explicit jobs so that we
can test additional python versions in a more flexible manner that isn't
directly tied to openstack.

Finally we remove experimental jobs because they are in check and gate
already.

Change-Id: I7edf5464f2c7f48cb9c68b9e3dbbb688bec8b292